Resolves issue #65 - treat classification byte as unsigned char

+++ src/main/java/com/github/mreutegg/laszip4j/laszip/LASreadItemCompressed_POINT10_v2.java
@@ -146,7 +146,7 @@ public class LASreadItemCompressed_POINT10_v2 extends LASreadItemCompressed {
                     m_classification[last_item.Classification] = dec.createSymbolModel(256);
                     dec.initSymbolModel(m_classification[last_item.Classification]);
                 }
-                last_item.Classification = (byte) dec.decodeSymbol(m_classification[last_item.Classification]);
+                last_item.Classification = (short) dec.decodeSymbol(m_classification[last_item.Classification]);
             }
 
             // decompress the scan_angle_rank ... if it has changed
